scripts/requirements.txt
# 爬蟲與網頁解析
requests
beautifulsoup4

# LLM API 串接 (看你想用哪家的，這邊先以 openai 和 google-genai 為例)
openai
google-genai

# 資料處理與輸出 report.csv
pandas